fix(Storybook): adds shell for component style addon refactor
